Move relayes to starknet main file
